Renaming sink to processor
