+++ src/logger/src/processor/consumer.rs
