+++ src/prelude/Cargo.toml
@@ -19,7 +19,7 @@ publish = true
 crate-type = ["cdylib", "rlib"]
 
 [dependencies]
-enso-shapely  = { version = "0.1.1" , path = "../shapely/impl" }
+enso-shapely = { version = "0.1.2" , path = "../shapely/impl" }
 
 boolinator   = "2.4.0"
 derivative   = "1.0.3"
